mononoke: Remove hook manager factory
Summary:
The hook manager factory was used in a couple of places to make it
simpler to create a `HookManager`. However, `HookManager` is now a facet and
can therefore be created using the facet factory methods.

Remove the hook manager factory code and move its logic into `repo_factory`.
Further, update `mononoke_api::Repo`'s `new_test_common` method to create a
`HookManager` using `TestRepoFactory`. In the future, we should get rid of the
custom test methods inside of `Repo` and rely on `TestRepoFactory` instead.

Reviewed By: markbt

Differential Revision: D38352454

fbshipit-source-id: aa127ad3408a049576dafe6e21bc65f50203ead2