mononoke: Convert mononoke_api::Repo to a container
Summary:
Now that all of the components of `Repo` are facets, the struct can be
converted to a container. This means that the struct can be built using a
`RepoFactory` rather than calling `Repo::new()`.

In later diffs, we can remove the helper methods (e.g. `hook_manager()`,
`ephemeral_store()` etc) and update callers to import traits instead.

Reviewed By: yancouto

Differential Revision: D38349354

fbshipit-source-id: bd9ae8b3d1e6aa1d0230306860fcff852161fad0