mononoke: Move skiplist env into repo_factory
Summary:
`mononoke_api::Repo` has logic to disable skiplists if there
`skiplist_enabled` isn't set in `MononokeEnvironment`. In order to migrate
`Repo` to be a container, this logic needs to live inside of the repo_factory.

Update the code to create a skiplist index to check the same environment
setting and disable skiplists if it's not set.

Reviewed By: yancouto

Differential Revision: D38349355

fbshipit-source-id: 0550607506b91e0a42a63bcec5995c20f12b03a0