prjfs: properly handle directory hierarchy replaced by a file
Summary:
When walking up the path that has been detected as not being present on disk,
EdenFS would assume that the returned path is a directory, but this assumption
doesn't hold true if the directory has been replaced by a file. Thus, instead
of walking up until finding a file/directory that is present on disk, the code
needs to walk up until a directory is found.

Reviewed By: kmancini

Differential Revision: D37927926

fbshipit-source-id: 5da55f503f4df8158c61435a1afa2cdd3f24d997