prjfs: properly handle renaming of directories
Summary:
Renaming of directories is handled in the same way that a directory was removed
and another one was created in a different location. Internally, EdenFS will
query the filesystem for whether the renamed directory exist, and remove the
entire inode hierarchy, at the same time, all the files in the destination
directory are added recursively in the inode hierarchy.

This model however breaks down if a new folder is then added on disk at the
location of the old directory. When that happens prior to the directory removal
described above, EdenFS will not removed the directory hierarchy, but it will
treat it as a directory creation, and will thus try to add all the on-disk
childrens of the created directory to the inode hierarchy. On completed, the
TreeInode corresponding to this directory will contain the old entries of the
renamed directory, thus a mismatch with what's on disk.

To solve this, we also need to make sure that all the entries in the TreeInode
are also consulted and recursed into. This will naturally remove the
files/directories that are no longer present on disk.

Reviewed By: kmancini

Differential Revision: D37764806

fbshipit-source-id: 249e2028cf419f837659d6813803b7a03c2a0163