Create Hg ActivityBuffer
Summary: This diff is part of a stack of diffs (D38276534 (https://github.com/facebookexperimental/eden/commit/464b402226505bf6a2285ccc814a8df89263e6d1), D38282758, D38287560, D38288599) for adding retroactive tracing support for the eden trace hg command. This diff specifically performs the logic within EdenFS to create an ActivityBuffer in each HgQueuedBackingStore and subscribe to the tracebus of this BackingStore to read and store hg events as they occur. In later diffs we will expose these events in a thrift interface, and then display them in the eden tracing CLI.

Reviewed By: MichaelCuevas

Differential Revision: D38282758

fbshipit-source-id: ef5a18482639ac6359a73ea255fe1035f97d103d