repo_authorization: implement path ACL checks
Summary:
Implement path ACL checks for basic read accesses to the repo.  Attempts to access
a tree or file will fail if the user doesn't have access to that path.  Attempts
to access arbitrary trees or files by their IDs will also fail if the user does not
have full repo read access, as the paths for these accesses cannot be checked.

In this initial version, we do not prevent queries that would allow users to
find out the names of paths that they do not have access to: for example
listing all files in a directory would still work.

Reviewed By: RajivTS

Differential Revision: D38129042

fbshipit-source-id: 1b2b6e5881712628c6d979d329447afac2050626