workingcopy: add basic pendingchanges API to filesystem
Summary: This is the first diff to make the workingcopy abstraction more formalized. Starting simple by creating a new filesystem crate that will define the filesystem abstraction, starting with this pendingchanges API. Move over the PendingChangesResult struct definition to here too. In the next diffs I will move over and refactor the filesystem implementations to conform to this API. The key difference betweem APIs is that details such as TreeState become an implementation detail of the filesystem and not part of the API.

Reviewed By: quark-zju

Differential Revision: D38354701

fbshipit-source-id: 738885104ec88a09e48bf5dafb095dcb91ca8c13