integration: use subprocess.poll instead of subprocess.wait
Summary:
We're seeing random timeout in this test in between EdenFS being killed, and it
being restarted. Looking at the code, the main culprit appears to be the
unbounded subprocess.wait call which Python's documentation warns about:
https://docs.python.org/3/library/subprocess.html?highlight=subprocess#subprocess.Popen.wait:

> Note: This will deadlock when using stdout=PIPE or stderr=PIPE and the child
> process generates enough output to a pipe such that it blocks waiting for the
> OS pipe buffer to accept more data.

Since the _process object is created with subprocess.PIPE, this looks to be a
good candidate for what we're seeing.

Instead of using subprocess.wait, we can also use subprocess.poll combined with
util.poll_until for the same behavior, but without the risk of deadlock.

Reviewed By: fanzeyi

Differential Revision: D38395869

fbshipit-source-id: acbae227644d9265cd3826102fbf5309f2a404d5