mononoke: Remove hook_manager helper function
Summary:
Now that `mononoke_api::Repo` is a container, it's possible to import
traits to access its facets. Remove the `hook_manager` method from the struct
and update callers to import traits.

NOTE: This is only a small change, but I am submitting it separately as I think
it's a good template for bootcamp tasks to be based on.

Reviewed By: markbt

Differential Revision: D38388650

fbshipit-source-id: 1c2ee9e3fbf31420836ffd6de513eb1ce88c8917