Add pagination support for commit_file_diffs
Summary:
Adds an optional parameter diff_size_limit to the commit_file_diffs thrift API, that limits the total byte size of the returned diffs. If that limit is reached the API returns the diff so far and the names of the file pair it stopped at.

Also adds diff_size_limit option to the scs client.

Reviewed By: markbt

Differential Revision: D38352190

fbshipit-source-id: 419ce0cc14bc10325da7e2d63962b61f0fec22ce