revisionstore: reduce size of aux store to 250MB
Summary:
An entry in the aux store is a maximum of 114 bytes, 10GB of disk space thus
allow storing 94M aux data, which is several times higher than the number of
files in fbsource. If we assume files to be ~5KB in size on average (I have no
clue if that's the case, just pulling this out of nowhere), then the
indexedlogdatastore can store roughly 200k files. Aiming for 10x the number of
aux data gives us roughly 250MB, let's thus size the aux store to this.

Reviewed By: DurhamG

Differential Revision: D38446306

fbshipit-source-id: 627114051d4030ec4e3339a003ded941d5d5b9c7