mononoke: lfs: Use connection security checker
Summary:
The LFS server has its own logic for checking whether a connecting
identity is a trusted proxy. However, the other Mononoke services have
converged on using `ConnectionSecurityChecker`.

Update the LFS server code to use that struct and remove the old code. Further,
update the integration tests to remove the `--trusted-proxy-identity` option as
that is now ignored.

Reviewed By: markbt

Differential Revision: D38429609

fbshipit-source-id: 30e49bf8517b356e2d3e3f6379be4ec4b8e091db