Implement BackupSourceRepo as a Facet Container
Summary: As part of facetization of `RepoHandler`, its two core components need to converted to `facets` or `containers`. This diff introduces a new struct `BackupSourceRepo` that facetizes one of the two core components in `RepoHandler`. Currently, this struct is constructed manually and I will be adding factory-based construction in the follow-up diff.

Reviewed By: HarveyHunt

Differential Revision: D38351537

fbshipit-source-id: 71b94b3dd316cd6e47c389c931d3b487476c9919