Prevent unnecessary BackupSourceRepo initialization for non-backup repos
Summary: As part of `RepoHandlers` initialization, the `BackupSourceRepo` instance is allocated whenever the `RepoConfig` for a repo points to a source-repo for itself (e.g. for `fbsource_backup`, fbsource serves as the source repo). However, currently the configs are defined to have a source repo regardless of whether its a backup repo or not. This diff prevents the unnecesssary initialization. A later diff will fix the config once I can confirm no other dependencies.

Reviewed By: HarveyHunt

Differential Revision: D38396229

fbshipit-source-id: b275f8966803b5c35f57df688fbc9000898ac339