Implement PushRedirectorBase + TargetRepoDbs as Facet
Summary:
This diff introduces a new struct `PushRedirectorBase` that is akin to `IncompletePushRedirectorArgs` and will replace all the intermediary `PushRedirectors` in the future post factory-based construction. `PushRedirectorBase` depends on `TargetRepoDbs` struct which in turn has been converted into a facet.

All the structs used for wireproto RepoHandlers will be placed under a new crate `wireproto_handler`. `BackupSourceRepo` is also moved to this crate.

Reviewed By: HarveyHunt

Differential Revision: D38394933

fbshipit-source-id: 14b3122508583d62cbeae2f81ea38275e11637be