Remove n^2 path comparisons from Windows fsck
Summary:
The windows fsck logic had code to find the OverlayEntry for a given
child name. This iterated over the entire parent directory contents until it
found the entry. This would happen inside a loop over the potential children,
resulting in n^2 path comparisons as we rescanned the overlay children (O(n))
for every potential child (O(n)).

Let's precompute the hashmap and just use that.

Reviewed By: xavierd

Differential Revision: D37796955

fbshipit-source-id: 0a7cdc544cd9f3dd65e04f70ee6ea4f156b314f3