Recover Overlay from disk/scm for Windows fsck
Summary:
Previously fsck for Windows had a very limited number of things it could fix. It could remove Overlay entries that no longer match disk and add Overlay entries for files on disk that were missing from the Overlay. It couldn't fix incorrect modes, fix incorrect scm hashes, or recover files from source control that should be in the overlay.

This new fsck implementation handles all that. Even if the Overlay is completely missing, it should be able to rebuild a working version of it.

Reviewed By: xavierd

Differential Revision: D37692081

fbshipit-source-id: 9a92817172557d4a0bba14c2b2ac4d53c3cc4065