Modify the thrift interface for HookOutcome to allow returning multiple rejections
Summary: Right now the scs endpoint commit_run_hooks has a bug where it only keeps one of the multiple HookOutcomes that ChangesetContext::run_hooks returns for a single hook. In order to fix it we should change the thrift interface to allow commit_run_hooks to return a list of rejections and modify the usage of this endpoint in www accordingly.

Reviewed By: mitrandir77

Differential Revision: D38534032

fbshipit-source-id: c1003885bc7701b0657eba90bbeb58af70d3e46d