mononoke: lfs: Use RepoPermissionChecker
Summary:
The LFS server doesn't use the shared code for checking authorisation
of clients, as the LFS server's code pre-dates that. Remove the custom
authorisation checking code and replace it with the `RepoPermissionChecker`.

This also means that the custom command line arguments are removed, which I've
taken out of the integration tests.

NOTE: Whilst Mononoke code now uses `AuthorizationContext`, the LFS server
can't be ported to that yet as it doesn't build a `Metadata` structure from a
client's connection.

Reviewed By: yancouto

Differential Revision: D38499246

fbshipit-source-id: 1339f7fe65f0643786f3ceafe5dfbc35e825e184