workingcopy: Give Rust status exclusive ownership of TreeState
Summary:
The Arc<Mutex<TreeState>> has been a constant source of problems, e.g. deadlocks. Now that we're trying to build out a proper workingcopy abstraction we'd like to have true ownership of TreeState, which is very reasonable when we have this fully in Rust. Not having ownership will cause us to have lifetime scattered all over our API which I'm keen to avoid, even if it's only temporary.

It was noted in D20546904 (https://github.com/facebookexperimental/eden/commit/c1b8f86359991a3110b419f0899dd5619af9b971) that our hand was forced to make this Arc<Mutex<TreeState>> instead of Rc<RefCell<TreeState>> because of CPython's requirements to make Python classes Send. Given that enough of workingcopy is now in Rust that we don't need to re-enter Python, e.g. for ignore matching or unravelling pending changes, we can take exclusive ownership of the TreeState for the entirety of the Status command and give it back when are done.

Reviewed By: DurhamG

Differential Revision: D38458623

fbshipit-source-id: fe7a7bc6fea3b7db8936766a5a58bd5c93d9958c