stack-config-derive: allow use of Option type
Summary:
The issue is that, if there is a nested StackConfig struct wrapped in `Option<T>`, stack-config-derive macro does not handle it correctly.

In the previous version, I took the shortcut and calculated the type name by simply concatenating strings. So it was generating: `__StackConfig_OptOption<NestedConfig>` for type `Option<NestedConfig>`, which is clearly incorrect.

In this diff, I introduced a `ConfigLayerType` trait for calculating the type and as a result I can provide a blanket impl for `Option<T>` which allows the user of the macro to nest types.

----

hoping this diagram can make the purpose of stack-config clearer:

{F758914485}

Reviewed By: xavierd

Differential Revision: D38478922

fbshipit-source-id: 6d22c5df37e8347218300cc5dc283b1feffda01e