store: remove RocksDbLocalStore::getImmediateFuture
Summary:
Reading from RocksDB is extremely fast, to the point where the context switch
cost is significant. In large fan out workloads such as glob, the vast majority
of EdenFS time is spent context switching and enqueing into the ioPool executor
for every walked tree. This not only causes CPU usage to be high, it also slows
down globbing, and causes very large amount of RAM usage too.

Since reading for RocksDB is fast, we should prefer doing the read in the
current thread, at the expense of blocking the current executor during the read
operation. The other downside is that we're losing a fault point, but the cost
of that fault point is too high for the benefit, we should thus remove it.

Reviewed By: fanzeyi

Differential Revision: D38551406

fbshipit-source-id: a675b0d5713cd3ce17df87e6e78c6b33944a30e7