status: remove dependency on clidispatch from edenfs-client
Summary:
Because I want to add a WorkingCopy command to clidispatch it introduced a circular dependency: clidispatch <- workingcopy <- edenfs-client <- clidispatch.

io was already a separate crate so that was easy enough. The other solution was to add a new error to signal which status cases are unsupported and have clidispatch decide to interpret that as FallbackToPython, which makes sense.

Adding an explicit OperationNotSupported error actually highlights that we're treating errors inconsistently here. Instead of just using it for cases that are actually unsupported in Rust and we want to fallback to python, we're returning it for *some* organic errors and hoping for the best in python. In another diff I will revisit this to make this cleaner and ensure we only fallback to python in the unsupported cases.

Reviewed By: DurhamG

Differential Revision: D38616507

fbshipit-source-id: 6945f6dfa5fe03fa7546993a38eda1d049ae2cfc