mononoke: Use automock for RepoPermissionChecker
Summary:
In a later diff I'd like to check that certain methods of
`RepoPermissionChecker` are called a specific number of times. Using `mockall`
makes it much easier.

Update the `RepoPermissionChecker` trait to use `mockall::automock` so that a
mock implementation can be created later. I had to include lifetimes in the
`check_if_region_read_access_allowed` in order to get this to compile. Thanks
to markbt and farnz for helping with that!

I also renamed `AlwaysAllowMockRepoPermissionChecker` to
`AlwaysAllowRepoPermissionChecker` to make sure it isn't confused with the
`MockRepoPermissionChecker`.

Reviewed By: mitrandir77

Differential Revision: D38653668

fbshipit-source-id: 9e2be3a4a6f85e82d61abb64fbd8c47de0f74d9d