make setting up fb303 monitoring easier
Summary:
When I worked on Fb303Args I was mostly targetting the services like
monnoke_server  that use serve_forever_async to start the main function.

Little did I know about existance of `app.run()` helper and that it would be
use as a replacement for `helpers::block_execute()`. Unfortunately it doesn't
have full feature parity as block_execute did set up the fb303 stats collection
while the app.run does not.

To deal with that I'm introducing new function meant as a drop-in replacement
for `block_execute` and I migrate the existing new framework code that had to
set up the collection manually.

Reviewed By: yancouto

Differential Revision: D38584472

fbshipit-source-id: b76d94dddc804b1183ee35c23c67a2b122825aa9