testing: improve compatibility with buck test
Summary:
With `buck test` the entry point of `hg` is not a native executable, but a
shell script that sets up a bunch of environment variables then run the
`hg_rust` binary.

That requires special handling in a bunch of places. For example,
`debugruntest` (or `edenscm.testing.ext.hg`) is restrictive about `PATH`, which
breaks the buck hg wrapper script, since `basename` is not in `PATH`. Be aware
of this case and relax the `PATH` environment to make buck test pass.

Reviewed By: DurhamG

Differential Revision: D38718449

fbshipit-source-id: 7914c662ca4616ef19fe6846516b39a4870803f1