Introduce MononokeRepos for per-service instance repo-access
Summary: This diff introduces `MononokeRepos`, which will be a dynamically reloadable collection of repos associated with an instance of Mononoke service (e.g. `eden-api`, `scs`, `lfs`) or command. The API allows access to individual repos by name or by id and provides a mechanism to iterate the entire set of repos. The repos can also be atomically added / removed / updated without affecting the existing processes.

Reviewed By: markbt, mitrandir77

Differential Revision: D38491511

fbshipit-source-id: 525f551983c173dc241029d6ba51ecc2bc860ed8