store: speed up ObjectCache by 1.64x
Summary:
Globbing is dominated by the ObjectStore::getTree time, which in benchmark is
mostly dominated by the cost of searching in the ObjectCache. The
std::unordered_map is known to have sub-par performance characteristics and
folly::F14 is a well know much faster replacement for it. Ideally, we would use
a F14FastMap, but the ObjectCache relies on objects not moving, using a
F14NodeMap which is a drop-in replacement for a std::unordered_map still yields
a significant speedup. Future changes will try to change the ObjectCache so
objects can be moved freely.

Reviewed By: chadaustin

Differential Revision: D38664358

fbshipit-source-id: 47eec8d91f4c2c3912812c0ce4d630055a382dce