Fix symlink process for the new Windows Fsck
Summary:
Previously, the new Windows Fsck logic was seeing dmode_t::Unknown for symlink files, which it then stored in the overlay. This caused later hg status's to see the file as having been modified.

This diff makes TreeEntry report dtype_t::Regular for symlinks on Windows. Thanks to xavierd for pointing this out.

Reviewed By: xavierd

Differential Revision: D38793312

fbshipit-source-id: dfff9ddac7a6c02503c01ea83a4486bc2687722b