Catching a windows problem in buck stop and adding an ignore and a helpful message
Summary:
There is a man-machine race condition on windows where if we shut down *eden* (even gracefully), for a few minutes Windows responds with "Filesystem provider unavailable" for all queries into the unmounted repo. If the user tries to delete the repo with *eden rm*, they will receive an error:
   error stoping aux processes C:\open\fbsource: [Errno 22] Invalid argument: 'C:\\open\\fbsource\\fbcode\\.buckd\\pid'

If we ignore the error and assume buck IS down (there's not much we can do anyway since the filesystem is unmounted), *eden rm* succeeds.

This diff ignore the Invalid Argument error on Windows and gives the user a meaningful message to retry in a few minutes in case of failure.

See test plan for repro steps.

Reviewed By: xavierd

Differential Revision: D38763837

fbshipit-source-id: 3679d91b2b1e40fbfa6ddfca099dce788828ab55