dispatch: move some cli config logic onto the ui object
Summary:
Push the cli config overrides and option derived config value logic onto the ui object. The ui object remembers the necessary inputs so a complete config can be reconstructed without the dispatcher's help.

In particular, this will allow loading a "remote" repo's config when we only have the "local" repo's config in-hand. I'm trying to streamline the config loading, and that means Python will get handed the local repo's config, so it no longer has a "base" config. This change allows us to derive a base config given a repo config.

Reviewed By: DurhamG

Differential Revision: D38250239

fbshipit-source-id: cefca1623156f8c7a18f4f4e70e75f1ea64ce0e1