testing: report python crash without losing stdout
Summary:
When interpreting bash code, if an Python exception was raised, the previous
stdout can be lost, which can be annoying for debugging sometimes.

Change `sheval` used by the test runner to instead turn the exception to extra
output in stdout. Preserves previous stdout output that is helpful for
debugging.

Differential Revision: D38815653

fbshipit-source-id: d4f149ca67025f4c080ff7dba7038f59afb070cf