service: ensure background globs are running in the background
Summary:
With ImmediateFuture, computation is run eagerly for as long as immediate
values are available as this is what is for the most part desired from a
performance perspective. However, there are cases where the eager computation
is not desired: when a background computation is requested. This can happen
with a `eden prefetch --background`. To fix this, we can simply start the
computation in a SemiFuture which will force the ImmediateFuture to be lazy as
opposed to being eager.

Reviewed By: MichaelCuevas

Differential Revision: D38849161

fbshipit-source-id: a7a6a450085bc913755ffe98b046a29250148809