Adding progress bar to eden rm
Summary:
We would like to improve the UX of the *eden rm* command. The *eden rm* command seems to "hang" from the users' perspective when it is performing a large task such as deleting a large directory tree in a repo that has been crawled in for some or other reason. This diff introduces a couple of print outs to indicate what the command is doing, and is introducing a progress spinner for the *rmtree* command.

Since the *shutil.rmtree* command is a black box, we can take advantage of the internal recursive behavior and produce a spinner that spins as long as the command keeps executing.

Reviewed By: xavierd

Differential Revision: D38754866

fbshipit-source-id: d04b403baa8c2de00c11e3f4f851605afc099ba6