rebase: handle rev change on transaction close
Summary:
The rebase logic uses revision numbers in internal states like self.state and
self.destmap. With segmented changelog, the rev of a node might get re-assigned
on flush (transaction close). This can cause problems for rebase across
multiple transactions.

Ideally rebase gets updated to use nodes in its internal states. However that
turns out to be changing too many places and without a typechecker it's less
confident to do correctly. This diff introduces a shim that uses {node:node} as
source of truth but ducks as a {rev:rev} map to make rebase work with segmented
changelog's rev changes.

In the future, if we recreate the rebase algorithm, we should only use nodes.

Differential Revision: D38280514

fbshipit-source-id: 4ed5098c9e3d1fb5e81efb238b9471a14cd47c0a