allow delayed opening of sqlite db
Summary:
opening the localstore can delay starting up the EdenFS server. I don't think
we use the Sqlite LocalStore any where (though we use RockDB for the overlay).
But I am going to allow the RockDbLocalStore to use a delayed open pattern, so
might as well teach sqlite to work this way to.

Reviewed By: mshroyer, xavierd

Differential Revision: D38839387

fbshipit-source-id: 1496d421fe210cd8d409c6fcc4303ad79dab066a