make EdenFS responsive on startup
Summary:
eden can be really slow to start and during the startup if the thrift server
is not running the CLI will not know that EdenFS is running. So for example
`eden start` will try to start edenfs and then run into troubles trying to
get the eden lock file, `eden doctor` will tell users to start eden because it
is not running.

This can be very confusing. Locally for me usually what delays the start is
EdenFS opening RocksDB. Let's allow EdenFS to open it's thrift server before
rocksDB so that EdenFS can be more responsive during start.

Reviewed By: xavierd

Differential Revision: D38190050

fbshipit-source-id: 1dc5433747284e9c7e12e9214c6a805c9b74520e