config: finish sharing Rust config with Python
Summary:
Now the ConfigSet remembers which files it has loaded. In Python we skip (re)loading the repo config if the repo's hgrc was already loaded by the config in-hand.

To maintain the config validation warnings I exposed the validation directly so Python can call it when it skips the full reload. In the future these warnings should move into Rust.

I also tweaked fixconfig() to not expand relative "paths" if it doesn't know the repo root (i.e. at beginning of Python dispatch before repo has been determined). If running in the context of a repo, any relative paths will be expanded when the localrepo is initialized.

Reviewed By: quark-zju

Differential Revision: D37981059

fbshipit-source-id: c7f55f141e1d5a7e8c312606386909afe1e960b0