formatter: add a crate for the templates in Rust
Summary:
In our Python codebase we have the concept of the formatter, which can be used for specifying output based on the template. There are a few different formatters; depending on the template and some config options, some formatter is chosen over another.

The Python formatter does support pseudo-templates, namely the `plain`, `debug` and `json` templates. At the moment it's easy to support these kinds of pseudo-templates in Rust. Implementing these should reduce the fallbacks from pure Rust commands into Python, since some commands use almost exclusively some of those three templates.

Additionally, the Python formatter is somewhat clunky, and its true value lies in its dispatcher and the ability to allow specifying fields and values, which allows the true template engine to produce output accordingly.

This diff implements the idea of the dispatcher, but instead of having the clunky implementation for outputing using the templater, it adds an interface for making every item specify output depending on the type. At the moment, it only supports the plain formatter, which does not allow specifying fields. Another difference between this Rust formatter and the old Python formatter is that this Rust formatter is not tightly coupled with IO and file handling.

Differential Revision: D38621739

fbshipit-source-id: ed2129dbb571849cf94f34c6fa4f4fd95101ab0b