test: move tests to more broadly use the test executor
Summary:
With some parts of EdenFS relying on pushing background work to the server
executor, some of the tests are now hanging on unlimited `.get` on some
ImmediateFuture. To solve this, we should move these ImmediateFuture to be
executed on the server executor, and make sure that the `.get` are timed to
avoid hanging tests.

Reviewed By: kmancini

Differential Revision: D38214804

fbshipit-source-id: 49b44eff52ed339583062b83c2259d4d9e489b2f