utils: in debug mode ImmediateFuture are always SemiFuture
Summary:
Due to their eager nature, an ImmediateFuture will execute future continuation
immediately when constructed with either an immediate value, or a ready
SemiFuture. One of the drawback of doing that is that tests often manipulate
data that is already ready, and thus we may rarely execute the not-ready code
path. This can lead to use after free when the capture list of a continuation
is not properly set.

To avoid this, we can make ImmediateFuture always be a wrapper on a SemiFuture,
even when constructed from an immediate value. This way, the use after free
described above will always occur in debug/sanitized builds, and
AddressSanitizer will tell us when these occur.

Reviewed By: kmancini

Differential Revision: D37729622

fbshipit-source-id: bcd858f821cc39d63fb446b0891a067bee370a22