Make the no_questionable_filenames hook more actionable
Summary: I was looking at the no_questionable_filenames hook that blocks push. The error message just says that the filename is illegal. It's left up to the user to determine what exactly about the filename is illegal. IMHO it would be better if we can print out the stuff that's not allowed (e.g. spaces, hypen, etc.)

Reviewed By: YousefSalama

Differential Revision: D38904896

fbshipit-source-id: 06ea5399928b9082fa48a5759dd5346208b9e072