add support for symlinking copied files
Summary:
This is needed by AOSP folks - we failed to sync such config change last weekend!

We can't symlink copied file right now because mononoke doesn't know which file should be the symlink source (the original or the copy).

This diff changes behaviour so we choose the one that's the first in the mapping (in most cases that shouldn't matter anyway as the contents of the symlinked files will be exactly the same)

Differential Revision: D38904216

fbshipit-source-id: d0b498a63cbda6e8d7ca5fb598836fa064b946a7