clone: add metrics to both the Python and Rust commands
Summary:
The Python clone command did not have the proper telemetry for logging whether it was a Rust clone or not, and the type of clone being performed.

This also adds some information about whether EdenFS was started when performing a clone. This is necessary since at the moment EdenFS has a long startup delay.

Reviewed By: DurhamG

Differential Revision: D38885218

fbshipit-source-id: 5ce0f3be25c7c929c62c0e2a6bc28c7945a4f175