fsck: do not error out if working copy is messed up
Summary:
In the case where a bug in EdenFS prevented a placeholder from being removed
during update, EdenFS would be unable to remount the bad mount. This would
prevent engineers from saving their work which would be a big loss of time.

As a first step, let's log an error and continue, EdenFS will have a wonky `hg status`,
but that state can then be recovered from (removing the file on disk
with `hg revert` for instance). As a followup step, we can keep track of these
paths and manually call PrjDeleteFile right after the PrjfsChannel is
initialized.

Ideally, we shouldn't silently ignore these and instead add them to some
in-memory buffer than `eden doctor` can then consume, this is a bigger project
that can be tackled later.

Reviewed By: DurhamG

Differential Revision: D39029629

fbshipit-source-id: 774f0bf1fb125d5951558642926440022b9a076f