filectx: remove linkrev lookup during renamed()
Summary:
'hg log foo.txt' with fastlog checks for ctx[f].renamed() for the first
commit. This commit could have a potentially expensive linkrev lookup, which
causes a hang. Since linkrevs are now expensive, and treemanifest lookups are
now cheap, let's remove the linkrev optimization and just rely on the parent
node check.

This is the opposite of the situation 5 years ago when linkrevs were cheap and
manifest lookups were expensive.

Reviewed By: quark-zju

Differential Revision: D39103910

fbshipit-source-id: 2e048352fda1955bd1f9ab11fdc552b4c1c2ab16