add watchman log snippet to eden rage
Summary:
eden rage command collects debugging diagnostics and puts them in a paste. There is another command, `watchman-diag | pastry` which does the same for Watchman. Since EdenFS and Watchman interact together often, getting some Watchman information in eden rage would be helpful for the team when it comes to debugging user issues. Right now, we might have to ask a user to run `watchman-diag | pastry` after we look at their post. This back and forth causes some friction.

To avoid this, we add a new section to the rage report, "Watchman logs" with a link to a paste, similar to the "Verbose EdenFS logs" section.

Reviewed By: genevievehelsel

Differential Revision: D39027983

fbshipit-source-id: 9c5a1e8fc856966595fb6aaec54847befd7ffc08