identity: flesh out Rust identity lib
Summary:
Add an Identity struct that encapsulates the bits different between an hg and an sl repo.

I'm planning to pass the Identity object around where convenient, but I've got a feeling there are many ".hg" references in inconvenient places, so I'm also going to have a global Identity for convenience.

Reviewed By: DurhamG

Differential Revision: D39037640

fbshipit-source-id: 48dfb7ee6b940bed6632eaeb85c849edb79208a8