Add edenfsctl --checkout-dir flag
Summary:
Adds a flag to edenfsctl to let us specify a checkout directory outside our current working directory. This is necessary if we're going to use `buck2 run edenfsctl` as an `eden-dev` replacement, since buck requires us to run from the source directory, which may be different from the checkout we want edenfsctl to operate on.

With this change, we can approximate `eden-dev` like this:

```
% buck2 run edenfsctl -- --config-dir=$HOME/local/eden-dev-state --checkout-dir=$HOME/fbsource-dev trace fs
```

This is implemented by altering both the Rust and Python entry points to accept the new flag and change to the given directory.

Reviewed By: xavierd

Differential Revision: D38961247

fbshipit-source-id: 814306871ebf01f1cabab87c98e85adce68b4ac8