utils: add a makeNotReadyImmediateFuture
Summary:
There are cases where the code doesn't wish to execute the ImmediateFuture
continuation inline. For instance, if the code wants to spawn many CPU
intensive futures to be executed in background threads and then collected,
executing these inline would make the code slower.

One easy way to solve this would be to simply use SemiFuture and push them to
background threads, but the rest of the code would need to be ported to use
SemiFuture all over the place. The use of SemiFuture would also slow down the
callers that wish to execute the continuation inline, which isn't desired.

Thankfully, we can achieve both by using ImmediateFuture, the caller can just
decide if it wants the code to execute inline (a regular ImmediateFuture), or
not (by building a not ready ImmediateFuture).

Reviewed By: chadaustin

Differential Revision: D37330741

fbshipit-source-id: 911827f12725d7a0d64986737d02026386fe459e