service: force diff computation to the background
Summary:
The main benefit of the streaming API is that a client can stop it by simply
closing the connection and EdenFS will detect this and stop the diff.
Unfortunately, one drawback (or benefit, depending on what is desired) of
ImmediateFuture is that they execute eagerly as long as data is available and
can be computed.

The negative side effect is that the streaming API may diff the entire trees if
these are all in memory or present locally without giving a chance to the
Thrift client to stop it. To prevent this, we can just make sure that the
computation is done lazily by forcing an ImmediateFuture to be not ready, and
then adding the diff operation as a continuation.

Reviewed By: chadaustin

Differential Revision: D37330736

fbshipit-source-id: b21ab358cbb4d5b4de0450fe44a75d54a6de1e7f