log architecture on macOS to edenfs_events
Summary:
It's not possible to differentiate NFS m1 and intel users in edenfs_events
right now. We are currently rolling out to intel users and it would be good
to understand the real number of intel users on NFS vs fuse.

Let's log the cpu architecture so that we can differentiate them.

Reviewed By: mshroyer

Differential Revision: D39139996

fbshipit-source-id: 582804707a7816c41aa6c69594f7198d3ce5dfe1