manually add git as a dependency to the .deb
Summary:
For the `.deb`, we try to create the `Depends:` line by running
`dpkg-shlibdeps` on the `hg` executable, but this doesn't catch
the fact that we need to be able to shell out to `git`. This diff updates
the logic in the `Makefile` to add `git` to `Depends:` as a special case.

Reviewed By: DurhamG

Differential Revision: D39162642

fbshipit-source-id: 64c7decf91f241fe9fd81dbfd51a0399becb0237