utils: allow ImmediateFuture<T> to be constructed from a compatible type
Summary:
In some cases, it's desirable to be able to construct an ImmediateFuture<T>
from another type, like in case of inheritence where T is the base class. In
this case, a static_cast is required for the compiler to not error out, instead
we can slightly change ImmediateFuture constructor to allow this conversion.

A couple of manual ImmediateFuture constructors were no longer necessary and
thus removed as a result.

Reviewed By: genevievehelsel

Differential Revision: D39158019

fbshipit-source-id: eb3452ba33d941251d9b4efaee28f82fd2c2dbd3