configparser: sniff identity when loading config
Summary:
Now configparser::hg::load will sniff the identity from either the repo_path, if provided, or else the environment.

This only means that it will load the repo config from .sl instead of .hg as appropriate.

I also tweaked eagerepo to sniff the identity instead of assuming ".hg" to aid in testing.

Reviewed By: DurhamG

Differential Revision: D39040645

fbshipit-source-id: fce46d1537028634b644bfd3dc4de684b9f88410