config: fix duplicate config loading for .sl repos
Summary:
Now the ui stores an identity object. The identity starts off as the default environment identity, but when a repo is loaded we sniff the identity from the repo's dot dir.

We use the identity to fix the check as to whether the repo's hgrc has already been loaded.

Reviewed By: DurhamG

Differential Revision: D39105267

fbshipit-source-id: 9d726af125a3167df054841eb1f18a018e01e615